Replaced Mojo::Home with a config setting.
The default is /home/alex/farm/face, which is what I need, but can set
it in your config file. The solution using Mojo::Home only worked, if
you called the application from the directory in question. If you
start it using a service such as Monit, this no longer works and
reports / as your home.